fix(smolvlm): 修正模型加载的 dtype 参数使用错误 在3090上训练需要调整数据类型
将原本错误的 torch_dtype 替换为标准的 dtype 参数，适配最新的 huggingface 模型加载接口

+++ src/lerobot/policies/smolvla/smolvlm_with_expert.py
@@ -78,7 +78,7 @@ class SmolVLMWithExpertModel(nn.Module):
             self.vlm = AutoModelForImageTextToText.from_pretrained(
                 model_id,
                 device_map=device,
-                torch_dtype="bfloat16",
+                dtype=torch.float32,
                 low_cpu_mem_usage=True,
             )
             config = self.vlm.config
